fix: extract shared attention module, fix topbar badge mismatch
Extract needsAttention/acknowledgeAttention into shared attention.ts
so topbar and graph-view use the same logic. Topbar badge now only
shows for NEW errors (not historical). Prune stale session entries
from lastSeenErrors map on acknowledge.
